refactor: split disk exporter tests
Move the disk exporter tests into files that match the production helper boundaries.

Keep the change mechanical by preserving test bodies and only adjusting import blocks and shared helpers.
